Organize the runtime into bounded contexts
The runtime directory grouped files by purity, not by language. One
core folder held proof evaluation, workspace freshness, execution policy,
and exit codes side by side, and the shell mixed project loading, proof
running, workspace copies, and worker orchestration.

Split it into four contexts, each with its own core and shell:

  project/    load a config, discover and select Gate modules, describe
  proof/      run one Gate or one Proof and judge the outcome
  workspace/  isolated per-Gate copies and tree freshness
  run/        whole-project runs, execution policy, workers, exit codes

The architecture Gate now guards every */core/ directory with one
convention instead of a path list. The unused runtime/report.ts facade is
removed. Public exports and behaviour are unchanged.
