+++ packages/redproof/src/project/index.ts
@@ -1,0 +1,2 @@
+export * from './shell/loader.ts';
+export * from './shell/describe.ts';
