+++ packages/redproof/src/proof/core/evaluation.ts
