+++ packages/redproof/src/proof/core/outcome.ts
