+++ packages/redproof/src/proof/index.ts
@@ -1,0 +1,7 @@
+export type {
+  CompletedProofOutcome,
+  InfrastructureProofOutcome,
+  ProofInfrastructureError,
+  ProofOutcome,
+} from './core/outcome.ts';
+export { runGate, runProof } from './shell/runner.ts';
