+++ packages/redproof/src/reporter/terminal.ts
@@ -3,10 +3,10 @@ import { relative, resolve } from 'node:path';
 import type { CheckResult } from '../domain/check.ts';
 import type { Diagnostic } from '../domain/diagnostic.ts';
 import type { Rule } from '../domain/rule.ts';
-import type { GateDescription } from '../runtime/describe.ts';
+import type { GateDescription } from '../project/shell/describe.ts';
 import { buildGateReportModel, summarizeGateReports, type RunSummary } from './model.ts';
-import type { ProofOutcome } from '../runtime/core/proof-outcome.ts';
-import type { CheckProjectRun, GateRun } from '../runtime/shell/project-runner.ts';
+import type { ProofOutcome } from '../proof/core/outcome.ts';
+import type { CheckProjectRun, GateRun } from '../run/shell/project-runner.ts';
 
 const glyph = {
   pass: '✓',
