+++ packages/redproof/src/run/index.ts
@@ -1,0 +1,2 @@
+export type { CheckProjectRun, GateRun, ProveProjectRun } from './shell/project-runner.ts';
+export { checkProject, proveProject } from './shell/project-runner.ts';
