--- packages/redproof/src/runtime/report.ts
@@ -1,2 +1,0 @@
-// Compatibility facade. Reporting is owned by the reporter bounded context.
-export * from '../reporter/index.ts';
