+++ packages/redproof/src/workspace/index.ts
@@ -1,0 +1,1 @@
+export * from './shell/copy.ts';
